move `TableSnapshotProviderImpl` to space module.
